+++ magnet/cli/main.py
@@ -8,12 +8,15 @@ Example:
     >>> MagnetCLI.main(argv=['download', '--help'], _noexit=True)
     >>> MagnetCLI.main(argv=['download', 'helm', '--help'], _noexit=True)
     >>> MagnetCLI.main(argv=['evaluate', '--help'], _noexit=True)
+    >>> MagnetCLI.main(argv=['evaluate_legacy', '--help'], _noexit=True)
+    >>> MagnetCLI.main(argv=['evaluate_new', '--help'], _noexit=True)
     >>> # Test version works
     >>> MagnetCLI.main(argv=['--version'])
 """
 import kwconf
 from magnet.cli.download_cli import DownloadModalCLI
 from magnet.evaluation import EvaluationConfig
+from magnet.evaluation_new import NewEvaluationConfig
 from magnet import __version__
 
 
@@ -25,7 +28,12 @@ class MagnetCLI(kwconf.ModalCLI):
 
 
 MagnetCLI.register(DownloadModalCLI, command='download')
-MagnetCLI.register(EvaluationConfig, command='evaluate')
+# The historical evaluator gets an explicit migration name while `evaluate`
+# remains a compatibility alias until the new evaluator takes that name.
+MagnetCLI.register(
+    EvaluationConfig, command='evaluate_legacy', alias=['evaluate']
+)
+MagnetCLI.register(NewEvaluationConfig, command='evaluate_new')
 
 
 __cli__ = MagnetCLI